send sources first to improve perceived speed
